Ajout de la vue détaillée d'un produit, mise à jour des modèles et des templates pour afficher les informations supplémentaires, et amélioration du style CSS pour la présentation des détails du produit.

+++ apps/produits/models.py
@@ -109,7 +109,7 @@ class Produit(models.Model):
 
         for unite, valeur in conversion.items():
             if self.quantite >= valeur:
-                return self.quantite / valeur
+                return int(self.quantite / valeur)
             
         return None
